Moved socketio related handlers in handlers module
