+++ src/mtool/api/rest/auth.py
@@ -22,3 +22,20 @@ def token_required(f):
             return jsonify({'message': 'Token is invalid'}), 401
         return f(current_user, *args, **kwargs)
     return decorated
+
+
+# authenticate the client for websocket connection
+def check_authentication():
+    token = None
+    if 'x-access-token' in request.args:
+        token = request.args['x-access-token']
+    if not token:
+        return False
+    try:
+        data = jwt.decode(token, current_app.config['SECRET_KEY'])
+        current_user = connection_factory.get_current_user(data['_id'])
+        print("current user", current_user)
+        print("authorized user connected")
+    except BaseException:
+        return False
+    return True
